Fix fullscreen issue Add data updating to OsgView when entering fullscreen mode Remove copier connection with screen space quad to maintain visual correctness

+++ SurgSim/Graphics/RenderPass.cpp
@@ -100,7 +100,7 @@ void RenderPass::showColorTarget(int x, int y, int width, int height)
 	if (m_debugColor != nullptr)
 	{
 		m_debugColor->setLocation(x, y);
-		std::array<double, 2> dimensions = {width, height};
+		std::array<double, 2> dimensions = {static_cast<double>(width), static_cast<double>(height)};
 		m_debugColor->setSize(dimensions);
 		m_debugColor->setLocalActive(true);
 	}
@@ -124,7 +124,7 @@ void RenderPass::showDepthTarget(int x, int y, int width, int height)
 	if (m_debugDepth != nullptr)
 	{
 		m_debugDepth->setLocation(x, y);
-		std::array<double, 2> dimensions = {width, height};
+		std::array<double, 2> dimensions = {static_cast<double>(width), static_cast<double>(height)};
 		m_debugDepth->setSize(dimensions);
 		m_debugDepth->setLocalActive(true);
 	}
